+++ src/main/resources/com/atomist/rug/ts/cortex.json
