+++ src/test/scala/com/atomist/rug/rugdoc/TypeScriptClassGeneratorTest.scala
@@ -5,7 +5,7 @@ import com.atomist.rug.spi.SimpleTypeRegistry
 import com.atomist.rug.ts._
 import com.atomist.source.{ArtifactSource, FileArtifact, FileEditor}
 import org.scalatest.{FlatSpec, Matchers}
-import TypeGenerator.CortexJson
+import TypeGenerator._
 
 object TypeScriptClassGeneratorTest {
 
@@ -24,13 +24,6 @@ object TypeScriptClassGeneratorTest {
       override def edit(f: FileArtifact): FileArtifact =
       f.withContent(f.content.replace(TypeGenerationConfig.TestStubImports,
         """
-          |interface ProjectContext {}
-          |interface PathExpressionEngine {}
-          |interface TreeNode {}
-          |interface FormatInfo {}
-          |interface Addressed {
-          |   address(): string
-          |}
           |interface GraphNode {}
         """.stripMargin))
     }
@@ -54,7 +47,7 @@ class TypeScriptClassGeneratorTest extends FlatSpec with Matchers {
   // We don't need to test against project types like File
   it should "generate compilable typescript classes" in {
 
-    val types = new TypeGenerator().extract(CortexJson)
+    val types = new TypeGenerator(DefaultCortexDir, DefaultCortexStubDir).extract(CortexJson)
     val tr = new SimpleTypeRegistry(types)
 
     val td = new TypeScriptClassGenerator(tr)
